Separate `scheme_date_for` into its own module
To be mixed into to factory helpers,
spec helpers and cucumber feature helpers.
